Consolidate duplicated payments page formatters
Five formatters were copy-pasted (byte-identical) across the payments pages:
formatDateTime + formatLastSynced + getCurrencyFractionDigits + formatPriceAmount
(Catalog & Subscriptions), and the currency-amount formatter (Customers as
formatCurrencyAmount, Transactions as formatAmountValue). Extract them into a
shared `features/payments/format.ts` and import from the four pages.

The genuinely page-specific date formatters with different output —
TransactionsPage.formatEventDate, CustomersPage.formatDateTime ("03:30" vs
"3:30"), SubscriptionsPage.formatShortDate — are intentionally left in place to
avoid changing their rendering. Behavior-preserving consolidation.
