+++ packages/dashboard/src/features/payments/pages/SubscriptionsPage.tsx
@@ -19,6 +19,7 @@ import { usePaymentClientPagination } from '#features/payments/hooks/usePaymentC
 import { usePaymentCustomers } from '#features/payments/hooks/usePaymentCustomers';
 import type { CatalogPrice, CatalogProduct } from '#features/payments/types/catalog';
 import { usePaymentSubscriptions } from '#features/payments/hooks/usePaymentSubscriptions';
+import { formatDateTime, formatLastSynced, formatPriceAmount } from '#features/payments/format';
 import type {
   PaymentSubscription,
   PaymentSubscriptionItem,
@@ -43,26 +44,6 @@ const SUBSCRIPTION_ROW_GRID_TEMPLATE =
   '32px minmax(0, 1.3fr) minmax(0, 1fr) 100px minmax(0, 1.2fr) minmax(0, 0.75fr)';
 
 const SUBSCRIPTION_ITEM_GRID_TEMPLATE = 'minmax(0, 1.2fr) minmax(0, 1fr) minmax(0, 1fr) 100px';
-
-function formatDate(value: string | null) {
-  if (!value) {
-    return '-';
-  }
-
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) {
-    return value;
-  }
-
-  return new Intl.DateTimeFormat(undefined, {
-    dateStyle: 'medium',
-    timeStyle: 'short',
-  }).format(date);
-}
-
-function formatLastSynced(value: string | null) {
-  return value ? formatDate(value) : 'Never';
-}
 
 function formatShortDate(value: string | null) {
   if (!value) {
@@ -95,7 +76,7 @@ function getSubscriptionStatusDisplay(subscription: PaymentSubscription) {
     return {
       label: 'Cancelling',
       className: SUBSCRIPTION_STATUS_CLASSES.cancelling,
-      tooltip: `Cancels on ${formatDate(subscription.cancelAt)}`,
+      tooltip: `Cancels on ${formatDateTime(subscription.cancelAt)}`,
     };
   }
 
@@ -105,7 +86,7 @@ function getSubscriptionStatusDisplay(subscription: PaymentSubscription) {
     return {
       label: formatStatusLabel(subscription.status),
       className: SUBSCRIPTION_STATUS_CLASSES[subscription.status],
-      tooltip: canceledAt ? `Canceled on ${formatDate(canceledAt)}` : null,
+      tooltip: canceledAt ? `Canceled on ${formatDateTime(canceledAt)}` : null,
     };
   }
 
@@ -124,34 +105,6 @@ function formatPeriod(subscription: PaymentSubscription) {
   return `${formatShortDate(subscription.currentPeriodStart)} - ${formatShortDate(
     subscription.currentPeriodEnd
   )}`;
-}
-
-function getCurrencyFractionDigits(currency: string) {
-  return (
-    new Intl.NumberFormat(undefined, {
-      style: 'currency',
-      currency: currency.toUpperCase(),
-      currencyDisplay: 'code',
-    }).resolvedOptions().maximumFractionDigits ?? 2
-  );
-}
-
-function formatPriceAmount(price: CatalogPrice) {
-  const rawAmount =
-    price.unitAmount ?? (price.unitAmountDecimal ? Number(price.unitAmountDecimal) : null);
-
-  if (rawAmount === null || Number.isNaN(rawAmount)) {
-    return 'Custom';
-  }
-
-  const currency = price.currency.toUpperCase();
-  const fractionDigits = getCurrencyFractionDigits(currency);
-
-  return new Intl.NumberFormat(undefined, {
-    style: 'currency',
-    currency,
-    currencyDisplay: 'code',
-  }).format(rawAmount / 10 ** fractionDigits);
 }
 
 function getCustomerLabel(customer: PaymentCustomer | null, subscription: PaymentSubscription) {
